Authorization

+++ routes/campground.js
@@ -1,20 +1,10 @@
 const express = require("express");
 const router = express.Router();
 const asyncWrapper = require("../utility/asyncWrapper");
-const ExpressError = require("../utility/ExpressError");
-const Campground = require("../model/campground");
-const { campgroundSchema, reviewSchema } = require("../schemas.js");
-const { isLoggedIn } = require("../middleware");
 
-const validateCampground = (req, res, next) => {
-  const { error } = campgroundSchema.validate(req.body);
-  if (error) {
-    const msg = error.details.map((ele) => ele.message).join(",");
-    throw new ExpressError(msg, 400);
-  } else {
-    next();
-  }
-};
+const Campground = require("../model/campground");
+
+const { isLoggedIn, validateCampground, isAuthor } = require("../middleware");
 
 //index
 router.get(
@@ -39,6 +29,7 @@ router.post(
     req.body.campground.location = `${req.body.campground.locationC}, ${req.body.campground.locationP}`;
 
     const campgrounds = new Campground(req.body.campground);
+    campgrounds.author = req.user._id;
     await campgrounds.save();
     req.flash("success", "Successfully made a new campground");
     res.redirect("/campgrounds");
@@ -51,7 +42,15 @@ router.get(
 
   asyncWrapper(async (req, res) => {
     const { id } = req.params;
-    const campgrounds = await Campground.findById(id).populate("reviews");
+    const campgrounds = await Campground.findById(id)
+      .populate({
+        path: "reviews",
+        populate: {
+          path: "author",
+        },
+      })
+      .populate("author");
+
     if (!campgrounds) {
       req.flash("error", "Cannot find that Campground");
       return res.redirect("/campgrounds");
@@ -64,6 +63,7 @@ router.get(
 router.get(
   "/:id/edit",
   isLoggedIn,
+  isAuthor,
   asyncWrapper(async (req, res) => {
     const { id } = req.params;
     const campgrounds = await Campground.findById(id);
@@ -71,6 +71,7 @@ router.get(
       req.flash("error", "Cannot find that Campground");
       return res.redirect("/campgrounds");
     }
+
     const locArr = campgrounds.location.split(",");
     locArr[0] = locArr[0].trim();
     locArr[1] = locArr[1].trim();
@@ -82,20 +83,20 @@ router.get(
 router.patch(
   "/:id",
   isLoggedIn,
+  isAuthor,
   validateCampground,
   asyncWrapper(async (req, res) => {
-    const { id } = req.params;
     req.body.campground.location = `${req.body.campground.locationC}, ${req.body.campground.locationP}`;
 
-    const campgrounds = await Campground.findByIdAndUpdate(
+    const campground = await Campground.findByIdAndUpdate(
       id,
       { ...req.body.campground },
       {
         runValidators: true,
       }
     );
-    res.flash("success", "Successfully updated");
-    res.redirect(`/campgrounds/${campgrounds._id}`);
+    req.flash("success", "Successfully updated");
+    res.redirect(`/campgrounds/${campground._id}`);
   })
 );
 
@@ -103,8 +104,16 @@ router.patch(
 router.delete(
   "/:id",
   isLoggedIn,
+  isAuthor,
   asyncWrapper(async (req, res) => {
     const { id: sid } = req.params;
+
+    const campground = await Campground.findById(id);
+
+    if (!campground.author.equals(req.user._id)) {
+      req.flash("error", "No permission!");
+      return res.redirect(`/campgrounds/${id}`);
+    }
 
     await Campground.findByIdAndDelete(sid);
     req.flash("success", "Deleted campground");
